fix : add authToken to API requests
