+++ features/workSession/api/services/workSession/endWorkSession.ts
@@ -9,10 +9,11 @@ import { EndWorkSessionParams } from '../../../types';
  * @return {*}  {Promise<void>}
  */
 export const endWorkSession = async ({
+  authToken,
   userId,
   workSessionId,
 }: EndWorkSessionParams): Promise<void> => {
-  const res = await axiosBase().put(
+  const res = await axiosBase(authToken).put(
     getApiEndpointFull('endWorkSession', { userId, workSessionId }),
   );
   return res.data;
